feat(settings): start new enum JvmSettings #7000
To be able to have more control over JVM settings names,
avoid typos, maybe create lists of settings and so on,
this enum will provide the place to add any old and new
settings that are destined to be made at the JVM level.

Further extensions of this class include adding aliases
when renaming settings, adding predicates for validation
and offering injecting parameters into keys (as used with
the files subsystem).