feat(settings): let AliasConfigSource.getValue use patterned aliases
Now that we store patterns for the keys handed to getValue(), we check if the
key starts with the commonly known prefix. Otherwise we do not execute any seeks
to spare some CPU cycles. (Anything non-Dataverse related would not be in the list,
so the expensive regex matching would be executed.)

For any Dataverse related setting, we match on the static list, and, if not found,
seek via the pattern list. Constructing a list of old names makes the support for
multiple historic names for both static and patterned names.