chore(deps): switch to Payara 6.2022.1-Alpha3 and Jakarta EE 10 #8305
With the release of Payara 6.2022.1-Alpha3, the application server has
been transformed into _requiring_ a Jakarta EE 10 app deployment.

This commit changes the Payara version, which updates the BOM, which is
built for EE 10. It also removes the EE API version property to be
aligned with the EE 10 BOM.

Most of the EE 10 JARs aren't available on Maven Central yet, because
it has not been officially released yet. It is necessary to add the
Eclipse Staging Repository temporarily. Should be removed when all
set upstream.