feat(ct-base): make image names configurable and rename
Add new Maven properties to choose a different Java base image
and change the name of the target base image when people customize it.
Also changes the build arg for the Java base image name.

With this, the image name changes to follow the same convention
as the Java base image.