refactor(ct-base): enable JMX, AMX and tune monitoring levels
Instead of acivating the JMX settings by default, switch back to off.
Extending this with enabling AMX and with all the log levels set to HIGH,
this might cause a huge performance overhead.
Also lacking is a MicroProfile Metrics mapping to retrieve JMX data
via the /metrics endpoint for collection.