style(ct-base): make up base image name from tag and add default
With defaulting to develop, we rest on using any build of the image during
experimentation etc to go with a (local) develop tag. Removing
the Java version from the tag makes it easier to use and reflects
the nature of it. It aligns image builds with the release schema
of the actual application while still allowing for experiments
and having different sources of truth for released and develop code.