docs(ct-base): clarify support image tags #8932
Adding notes about the image tags produced by the community
for reuse in the community. Document final tagging
strategy, using the branch name (develop/main) instead of
the Java version or sth.

Reshape the automated builds and publishing part to be included
in the supported tags and build instructions section to reduce
text complexity and group matching parts together.