refactor(metadata): rename CodeMeta softwareVersion to codeVersion #7844
As the citation block already contains a compound field "software"
with both "softwareName" and "softwareVersion", meant to describe software
used to create the dataset, this name conflict must be resolved.

By renaming to "codeVersion", the semantic is not changed, as this metadata
block is about describing software deposits. As the termURI is explicitly
set to "schema.org/softwareVersion" it remains compatible with OAI-ORE and
other linked data usages. A future exporter for CodeMeta might require
special attention for this field.