Fix for unrelated bug found in QA
In the UI doing a file replace and, in the case where the mimetype does
not match the original, selecting delete could result in a broken upload
tab. This only occured with direct upload enabled and was due to the
delete call not reinitializing the upload tab for direct upload.