fix(mocks): add authority to mocked dataset
For a lot of places where files are read from filesystem or other places,
both authority and identifier of the dataset are required to build a
proper path.

When either of those is missing (=null), all sorts of NPEs will occur, as
paths cannot be built with null values. In reality, this will hardly ever
occur, as a missing identifier let alone authority is completely unrealistic.

Yet, some functions are not validating this data very well. Adding this
mock here means covering potential NPEs, but again: these are unlikely to happen
outside unit tests.