fix(storage): avoid NPE when building path to local storage dir
During unit testing it occured that sometimes parts of paths are null (there was a missing mock for authority).
Changing the code structure here to catch this anytime.

This hard to track cause of trouble was revealed when running tests with JDK 17, because since
JDK 16 all java.nio.file.Path may not contain null elements for the first component.

See also: https://stackoverflow.com/questions/68791709