fix(settings): switch oridinality of AliasConfigSource to 100 #7000
To enable the AliasConfigSource to provide settings that also have a
representation in a microprofile-config.properties file, it needs to have
at least the same ordinal as that source. Thus the ordinal shall not be
minimum integer, but 100. As the name is "Alias", it will likely to be
used first when other sources are present with this ordinal (like that
properties source).