feat(pid): add dataset version pid configuration code #4499
This commit adds a new scope and setting to the JvmSettings,
enabling the configuration of different modes for Dataset Version PIDs.
These modes are depicted in VersionPidMode. A test ensures the
parsability.

In addition, VersionPidMode also contains a fine grained option
to change the conduct of Dataverse collections and their datasets
for these PIDs.