feat(collections): add business logic to determine dataset version pid conduct #4499
This commit adds a public method DataverseServiceBean.wantsDatasetVersionPids()
that will determine how to deal with a dataset version (which belongs to a dataset
that lives within a collection) in terms of "should a PID be registered/updated?".

The background is: when a dataset is published, there will be the context of the
owning Dataverse collection. It's important to take into account the configured
conduct for the collection in the decision how to go ahead with a version's PID.