fix(pid): make EZID provider not login if not configured
This relates to #9486, but with the switch of the properties
to come from MPCONFIG, this is much easier to do.

We do create the provider object in any case, because otherwise the
stateless EJB mechanism will be very unhappy. We do not login
if there is no password and username given, so avoid the exceptions.
We only log a warning if only one of both is given (which is clearly an
error).