feat(pid): enable version PID generation when new version is created
Any new version that is created by executing the CreateDatasetVersionCommand
will need to have an identifier generated at that precise moment.
Depending on the activation status, the generation is triggered.
As the global PID service needs to retrieved from the context of the command engine,
the context is added as a parameter to the method.