Revert "feat(pid): enable version PID generation when new version is created"
This reverts commit 5288bef9b27a77167d2dc077d80facbc7854f650.

After investigation, the timing to wire the generation in like here is
ill advised. This command is executed when a *new* dataset is about to
be created and all of the details in the DatasetVersion are still
"null"! (No status yet, no version numbers, etc)

As we will never know a-priori if a version is going to be major or
minor (that depends on users' actions and someone making a choice when
hitting "publish"), there is no reason to pre-register a version PID
that might never see the light of day.

A version PID for a minor version will carry the PID of it's major
revision, as it is a simple update. This will require changing the
logic some more.