refactor(pid): replace exception for GlobalIdServiceBean.createIdentifier()
So far, the method to create identifiers was defined in the interface to be
throwing any Exception or other Throwable it likes to throw. This is a dark
antipattern and should be avoided at all cost.

With this change, the interface defines the method to throw IOExceptions,
which is a much better description of what kind of errors to expect.
Per definition, an I/O exception "signals that an I/O exception of some sort has occurred.
This class is the general class of exceptions produced by failed or interrupted I/O operations."
Which is exactly what happens if for some reason the communication between us
and the PID provider goes sideways.

It also adds Javadoc to the interface. Up until now, it was no where documented
what the method will or has to return and what this method should and should not do
when implemented for a provider.