refactor(cmd): add NotImplementedException
By adding a NotImplementedException, a subclass of UnsupportedOperationException,
we can describe methods not yet implemented. This status might change in the future (or not).

The important part about this introduction: this runtime exception is flagged
as an "application exception", which makes the EJB exception inspection not handle it
as a system exception (which is the default), resulting in rolling back transactions.

The idea is to enable handling perfectly fine situations within the command engine when
some other component of Dataverse - e.g. the EJB of a PID provider - throws this exception.
This way we can catch the exception and deal e.g. with dataset unlocks.