feat(pid): restructure version identifier related methods in GlobalIdServiceBean
With the addition of NotImplementedException we can better express what's going on without
rollback exceptions in EJB.

The methods that will reach out to some provider now throw checked IOExceptions,
as the communication might go sideways and the command engine needs to act accordingly.
(This is the same behaviour now as for the normal object methods)

Also extending JavaDoc descriptions and expectations a bit.