feat(pid,cmd): create and publish version identifiers on dataset publication
With this commit, a first draft to create and release a version identifier
from the command engine is added. FinalizeDatasetPublicationCommand is used
after someone hit publish in the UI and currently takes care of the creation
and publishing for dataset identifiers and files, now extended to look after
version PIDs as well.