feat(ct): enable bootstrapping of instances
With adding bootstrap.sh, a Dataverse instance may now be
initialized. Using the changed setup scripts, we can reach
out to a Dataverse container and set it up.

By using a concept of personas, it is easy to add and select
other initializations as necessary. This will be extended in
the future.