feat(auth): pass OAuth2/OIDC state parameter to provider
When the client is returning from the provider to us, carrying along
the authorization code we need to retrieve user details, we also
receive again the state. The state was generated and sent by us,
and will not be altered by the provider, which makes it perfect to
identify the original request we built before sending the client to
the provider.

Passing this state to the provider enables the provider to reuse this
information. This is crucial to enable PKCE support, as we need to
remember which secret code we sent to the provider - otherwise we
will not be able to verify the authz code.

Tests have been adapted.