build,test: reconfigure JaCoCo test coverage measuring
With the addition of integration tests we need to create a combined report
of unit tests and integration tests (for now still keeping API tests out
of the loop for normal cases).

This commit reconfigures the JaCoCo plugin with multiple executions at
certain times to execute the measurements and merge the results together.