feat(oidc): replace map-based cache of PKCE verifiers
Instead of using an unlimited growth Map of verifiers
as a cache, we now will use a real evicting cache implementation
to limit the size and age of entries.

This will limit resource waste and mitigate an unlikely
but present attack vector of pumping up the cache to DDoS us.