build(spi): add version suffix to enable snapshots
By default, we will release an SPI package when new code is going
to be merged into the develop branch. So in the develop branch
always lives a non-snapshotted version of the SPI package POM.

To enable snapshots for testing and experimentation, we want to
be able to add some string including "-SNAPSHOT" at the end.
This is the marker for Maven to send a package into a snapshot
repository (and there is no other way to do this).

By adding this property which defaults to empty, we can override
things in CI from commandline and actually release snapshots
on terms we define in CI.