fix(ct): enable sane default for upload storage location in containers
The default from microprofile-config.properties does NOT work,
as the location must already be resolvable while the servlet is
being initialized - the app shipped defaults file is not yet read
at this point.

This is similar to the database options, which must be set using one
of the other Payara included config sources. (Non-easily resolvable
timing issue). The solution for containers is to add an env var to
the docker file, which can be overriden by any env var from compose
or K8s etc. (Problem is the high ordinal of the env source though)