fix(ct,conf): switch to different approach to default upload location
Instead of trying to provide a default using STORAGE_DIR env var from
microprofile-config.properties as before, using this env var reference
in glassfish-web.xml directly now. By defaulting to "." if not present
(as in classic installations), it is fully equivalent to the former
hardcoded default value.

Providing a synced variant of it in microprofile-config.properties
and leaving a hint about the pitfalls, we can reuse the setting for
other purposes within the codebase as well (and expect the same
behaviour because same defaults).