backward compatible SWORD upload on Payara 6 #8305
As of Payara 6, you can no longer send
"Content-Disposition: filename=example.zip"

Instead you must send
"Content-Disposition: attachment; filename=example.zip"

For backward compatibility, we are adding "attachment; "
if it's missing.

An alternative would be to force this backward
incompatibility on our users. Originally we did this
in 89182c1 but we reverted it in 83c55a9