add no-op editFileMetadata() to FilePage to prevent errors #8305
As of Payara 6/JSF 4/Jakarta EE 10, this syntax doesn't work:

action="#{bean[editFileAction]()}"

(Or it logs errors in server.log at the very least.)

So we switch to this syntax:

action=#{bean.editFileMetadata()}"

This means we have to make sure the method exists on all the beans,
though, or we get errors. (It's unclear why we weren't seeing errors
before!) So we add the method to FilePage, which gets passed as bean.

We also remove the now unused "editFileAction".