feat(mail): add sane defaults for mail jvm settings at app level #7424
We only default to no authentication. We still require people to
configure an SMTP host, only in containers we do default to "smtp"
as a hostname for that (see our compose file).

Username/password cannot have a default and all other special settings
should not be done here. These are highly setup specific.