upgrade Sphinx, add Markdown support #10101 #10111
We don't seem to be using intersphinx and I got an error about it so i
simply removed it.

I also got rid of the Python 3.10 vs 3.11 divide. It's no longer
relevant and some older servers were holding us back. Everyone should
use the same (latest!) Sphinx.

I included a tiny conclusion.md file as a demo, but it can be removed,
of course.