fix(ct): make base image comply with OpenShift file permissions
To enable the user with a random, arbitrary UID to write into the overlay filesystem, we need to set proper file permissions.
This should not affect users on Docker or other K8s distributions, as the security is more lenient there.

It is not ideal to write into overlayfs, as it impacts performance and may lead to unintended side effects. This is a workaround to at least get going.

 See https://docs.openshift.com/container-platform/4.14/openshift_images/create-images.html#use-uid_create-images for a detailed reference