docs(ct): remove configuration trigger step in IntelliJ
Depending on the attachment configuration, the run configuration waits in blocking mode for more output from the container logs. The application would never be deployed, as the wait is indefinite.

We need to run the compose step and the deploy step on their own. One appears in the services tab, the other in the run tab.