cosmetic additions to globus-api.rst
A few cosmetic/trivial additions to the guide, if you want them. 
Things like the dataset id query parameters missing from the curl example lines... that are really obvious, and it may be safe to assume that only serious developers are going to read this guide... so, up to you.