remove intro page and put its content in index and code pages #10531
This is a bit of an experiment discussed with @vaidap and
@DieuwertjeBloemen where we are trying to put more content on
the index page rather than a table of contents.

Because we have a subpage (code.md) we still need to put it in
a toctree (or it will be an orphan), so I hid it the best I could
within the flow of the text.