build,doc(parent): add special base image versioning
As long as we are developing the next version of Dataverse, we don't want the develop branch to fight the container images updated by the maintenance workflow for the past releases.

To avoid this, we simply use the next release version for the base image we use during development.

It requires updating a property during the release process twice, as we need to set the version to the "main" version property ($revision) once we aim for creating the tag.

This way, when using the release tag in forks or other ways as a foundation, people end up with the right base image.

After the tag is done, we need to prepare using the virtual next version within the develop branch again.