build(ct,parent): adapt base image POM to new tagging policy
We will use a long tag with Payara and Java version in it during development. This way we auto-update images on dev's PCs on these breaking changes.

For any releases, we will use short tags using only the version and the OS flavor.

These rolling tags will be accompanied by the revisioned tags. These extend the base name with the suffix as before, but only for releases, not the development branch. As the development branch is a rolling target, immutable tags make little sense.

This commit also adds the necessary backporting patches to older releases for the parent module to define the base.image.version as $revision, as described in the extended release process.