move updateDatasetUser past the file updates
it apparently causes new datafiles, which don't have create dates to
start to be persisted, cause a failure with an sql insert to the dataset
table with a null createdate. Moving it back to this location assures
the datafiles have create dates and avoids the exception. It's not clear
to me why trying to get the authenticatedUser in the updateDatasetUser()
call causes this.