fix JsonPrinter and failing tests #10519
JsonPrinter was edited in #10764 and #11066 which were merged five and
two weeks ago, respectively. This caused merge conflicts which should
be resolved correctly now. In this commit, we are starting with
JsonPrinter as of the most recent "develop" branch commit (7fdb21a)
and adding in our DatasetType changes.

PR #10764 resulted in the number of metadata fields being reduced from
80 to 35 and this commit make the same change in an assertion.