stop loading Codemeta in Docker, add setDisplayOnCreate API #10519
In Docker, we were testing with the Codemeta block. This made for a nice
real-world story of creating a "software" dataset type and using it with
the Codemeta block.

The Codemeta block also has the advantage of having some fields that are
set to displayOnCreate, which helped us make assertions that the code is
working properly. It's the only non-citation block with fields set to
displayOnCreate.

However, Jenkins doesn't have the Codemeta block, meaning that tests are
failing. Also, we aren't ready to promote the Codemeta block to be
shipped with Dataverse because a new version is out:
https://github.com/IQSS/dataverse/issues/10859

So, we are switching from Codemeta to the Astrophysics block. We create
an "instrument" dataset type. Like all non-citation blocks (except for
Codemeta), there are no fields that are set to displayOnCreate=true.
Therefore, we added an API for this.