test(mail): update test SSL certs for maildev to be RSA 2028 again
It is not necessary to use 4096 bit, as the error was about the wrong common name in the cert.

Detailed explanation:
Before, the cert presented to Jakarta Mail neither contained a commonName=localhost nor subjectAltName=localhost.
This worked before with Payara 6.2024.6 because Eclipse Angus Mail 1.0.0 was still in use (see list of managed dependencies on the 6.2024.6 BOM).
Certificate host identity check was introduced as default=on with Eclipse Angus Mail 1.1.0 (see their release notes).
Now that Payara 6.2025.2 is in use on develop, we upgraded to Eclipse Angus Mail 2.3.0, making the error appear.