build(ct): add an app image version Maven property
We will need it to have the version number ready for the maintenance job.
At the same time, leaving a note that we will _not_ use it for the default tag we use during development.

If we'd use the version there, too, it would break many scripts that now can rely on the single rolling tag of "unstable", as we'd need to change the tags everywhere for a new version.