feat(migrate): add Dataverse database migration module with (optional) Docker and Flyway
This commit introduces a new Maven module, `dataverse-migrate-db`, for database migration. It utilizes Docker for setting up a PostgreSQL environment and Flyway for managing schema migrations. The configuration supports customizable properties and includes execution phases for setting up, dumping, migrating, and cleaning up the database.

The setup of Postgres in Docker can be skipped to run the migrations against some other remote database. It is up to the user to ensure no harm comes from that.