feat(migrate,ct): update dependency versions and relocate configuration
Moved `flyway.version` and `postgresql.server.version` to the `dataverse-parent` POM for better configuration centralization. Removed their redundant declarations from the root POM to maintain consistency. Enables reusing the values in the new `dataverse-migrate-db` module

Downgraded the PostgreSQL server version from 17 to the one specified in the installation guide (16) to be in sync with the official recommendations.