chore,ci(ct): update Renovate schedule to use cron syntax
Replaced the fixed Sunday 6:00 AM schedule with a cron expression as recommended by the docs. This maintains the behavior of creating PRs only on Sundays, but hopefully works correctly now.

Looking at the logs within the Mend Renovate Board, the general scanning etc works but the rule seems to not allow execution of PR creation for some reason. Educated guess: it's not 6 am when the scheduler runs and the logic is to exclusive.