style(ct): use app.image.version for configbaker OCI metadata tags
To have proper references also for future versions, we cannot just use project.version. (As seen with the app image Dockerfile, just transferring the method)

Needs a workaround for fabric8io/docker-maven-plugin#1865:
Included APP_IMAGE_VERSION in pom.xml and Dockerfile to address version handling issues with the plugin. This ensures proper image versioning and compatibility in the build process.