Squashing the commits in the branch:
no init in custom page

Hides the keycloak auth. option from the login page.

added a message to the login page about the auth. methods that are no longer supported.

removing the "if you believe this is an error" that Dataverse attaches to every error message in the UI.

A typo!