fix: add datasetType to JSON serialization of DatasetVersion
THe reason to add this here is that some client tools might request
specific versions of a dataset, but need the information about the type,
too. For example some client might want to display a specific version of
a software instead of a dataset differently.

It is debatable if this should be included here, as so far there is no
consensus yet wether or not the type of a dataset should be allowed to
change between versions of it. On the other hand, the global ID should
also never change, but it is included anyway for the sake of getting
back to the dataset via the API if necessary.

This should be revisited as a JSON API design issue at a later point,
especially as when creating a dataset version, the type is at the root
level and the version itself is at a sublevel, rendering the API
inconsistent from the start.