add to the enum many archive-related settings #11639
Attempts to set these settings were failing with a 400
and "The name of the setting is invalid." By adding them
to the enum, they pass validation and are able to work again.

These settings are really subsettings of :ArchiverSettings.
(See installation/config.rst.)

The exception is :BagGeneratorThreads, which is not a
subsetting of :ArchiverSettings. Rather, it is a related
setting used by multiple archivers.

:DRSArchiverConfig was added to the enum. DRS is a system
specific to Harvard which is why we don't document the setting
in the guides, but we plan to mention it in the release notes.
See also https://github.com/IQSS/dataverse.harvard.edu/issues/177