delete /api/admin/test but add some TODOs #11760
Enough functionality has been added in #11760 that it
should be save to delete all of /api/admin/test (TestApi.java)
because it was only testing external tools.

However, if we incorporate this change into that PR we should
do the following:

- fix the backend to not return a toolUrl when it shouldn't
(see testFileLevelToolWithAuxFileReq)
- consider how we aren't exercising findDatasetToolsByType
and findFileToolsByType anymore. They are used by JSF.
- consider how we don't return the following fields when
getting the toolUrl: scope, types, contentType
- remove commented code that's just there for review